Fix crash on WebStorm
This was due to an undeclared dependency on Groovy.
This has been replaced by a dependency on Javaluator.
This requires the plugin to take on a less permissive License.
This does not mean that users can't use the plugin for proprietary software, but that the *plugin itself* can't be redistributed in such a way that restricts access to the source code or ability to modify.